Moved functions into appropriate files
